requirements.txt
matplotlib==3.10.0
numpy==2.2.0
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.0
streamlit==1.41.1
tqdm==4.67.1
torch==2.5.1
torchvision==0.20.1
transformers==4.47.0